+++ src/main/java/net/sf/jabref/event/EntryChangedEvent.java
@@ -1,0 +1,18 @@
+package net.sf.jabref.event;
+
+import net.sf.jabref.model.entry.BibEntry;
+
+/**
+ * <code>EntryChangedEvent</code> is fired when a <code>BibEntry</code> has been changed.
+ */
+
+public class EntryChangedEvent extends EntryEvent {
+
+    /**
+     * @param bibEntry <code>BibEntry</code> object the changes were applied on.
+     */
+    public EntryChangedEvent(BibEntry bibEntry) {
+        super(bibEntry);
+    }
+
+}