+++ src/main/java/net/sf/jabref/model/entry/BibEntry.java
@@ -15,10 +15,6 @@
 */
 package net.sf.jabref.model.entry;
 
-import java.beans.PropertyChangeEvent;
-import java.beans.PropertyVetoException;
-import java.beans.VetoableChangeListener;
-import java.beans.VetoableChangeSupport;
 import java.text.DateFormat;
 import java.text.FieldPosition;
 import java.text.ParseException;
@@ -36,9 +32,11 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.TreeSet;
 
+import net.sf.jabref.event.FieldChangedEvent;
 import net.sf.jabref.model.database.BibDatabase;
 
 import com.google.common.base.Strings;
+import com.google.common.eventbus.EventBus;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
@@ -54,8 +52,6 @@ public class BibEntry implements Cloneable {
     private String type;
     private Map<String, String> fields = new HashMap<>();
 
-    private final VetoableChangeSupport changeSupport = new VetoableChangeSupport(this);
-
     // Search and grouping status is stored in boolean fields for quick reference:
     private boolean searchHit;
     private boolean groupHit;
@@ -70,9 +66,13 @@ public class BibEntry implements Cloneable {
     private boolean changed;
 
 
+    private final EventBus eventBus = new EventBus();
+
+
     /**
      * Constructs a new BibEntry. The internal ID is set to IdGenerator.next()
      */
+
     public BibEntry() {
         this(IdGenerator.next());
     }
@@ -108,12 +108,7 @@ public class BibEntry implements Cloneable {
     public void setId(String id) {
         Objects.requireNonNull(id, "Every BibEntry must have an ID");
 
-        try {
-            firePropertyChangedEvent(BibEntry.ID_FIELD, this.id, id);
-        } catch (PropertyVetoException pv) {
-            throw new IllegalStateException("Couldn't change ID: " + pv);
-        }
-
+        eventBus.post(new FieldChangedEvent(this, BibEntry.ID_FIELD, id));
         this.id = id;
         changed = true;
     }
@@ -167,18 +162,12 @@ public class BibEntry implements Cloneable {
             newType = type;
         }
 
-        String oldType = this.type;
-
-        try {
-            // We set the type before throwing the changeEvent, to enable
-            // the change listener to access the new value if the change
-            // sets off a change in database sorting etc.
-            this.type = newType.toLowerCase(Locale.ENGLISH);
-            changed = true;
-            firePropertyChangedEvent(TYPE_HEADER, oldType, newType);
-        } catch (PropertyVetoException pve) {
-            LOGGER.warn(pve);
-        }
+        // We set the type before throwing the changeEvent, to enable
+        // the change listener to access the new value if the change
+        // sets off a change in database sorting etc.
+        this.type = newType.toLowerCase(Locale.ENGLISH);
+        changed = true;
+        eventBus.post(new FieldChangedEvent(this, TYPE_HEADER, newType));
     }
 
     /**
@@ -365,19 +354,8 @@ public class BibEntry implements Cloneable {
 
         changed = true;
 
-        String oldValue = fields.get(fieldName);
-        try {
-            // We set the field before throwing the changeEvent, to enable
-            // the change listener to access the new value if the change
-            // sets off a change in database sorting etc.
-            fields.put(fieldName, value);
-            firePropertyChangedEvent(fieldName, oldValue, value);
-        } catch (PropertyVetoException pve) {
-            // Since we have already made the change, we must undo it since
-            // the change was rejected:
-            fields.put(fieldName, oldValue);
-            throw new IllegalArgumentException("Change rejected: " + pve);
-        }
+        fields.put(fieldName, value);
+        eventBus.post(new FieldChangedEvent(this, fieldName, value));
     }
 
     /**
@@ -394,14 +372,8 @@ public class BibEntry implements Cloneable {
         if (BibEntry.ID_FIELD.equals(fieldName)) {
             throw new IllegalArgumentException("The field name '" + name + "' is reserved");
         }
-        Object oldValue = fields.get(fieldName);
         fields.remove(fieldName);
-        try {
-            firePropertyChangedEvent(fieldName, oldValue, null);
-        } catch (PropertyVetoException pve) {
-            throw new IllegalArgumentException("Change rejected: " + pve);
-        }
-
+        eventBus.post(new FieldChangedEvent(this, fieldName, null));
     }
 
     /**
@@ -445,27 +417,6 @@ public class BibEntry implements Cloneable {
             }
         }
         return false;
-    }
-
-    private void firePropertyChangedEvent(String fieldName, Object oldValue, Object newValue)
-            throws PropertyVetoException {
-        changeSupport.fireVetoableChange(new PropertyChangeEvent(this, fieldName, oldValue, newValue));
-    }
-
-    /**
-     * Adds a VetoableChangeListener, which is notified of field
-     * changes. This is useful for an object that needs to update
-     * itself each time a field changes.
-     */
-    public void addPropertyChangeListener(VetoableChangeListener listener) {
-        changeSupport.addVetoableChangeListener(listener);
-    }
-
-    /**
-     * Removes a property listener.
-     */
-    public void removePropertyChangeListener(VetoableChangeListener listener) {
-        changeSupport.removeVetoableChangeListener(listener);
     }
 
     /**
@@ -651,4 +602,13 @@ public class BibEntry implements Cloneable {
     public int hashCode() {
         return Objects.hash(type, fields);
     }
+
+    public void registerListener(Object object) {
+        this.eventBus.register(object);
+    }
+
+    public void unregisterListener(Object object) {
+        this.eventBus.unregister(object);
+    }
+
 }